# api/control-plane/swagger.yaml
basePath: /
definitions:
  adminapi.FieldOption:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
  adminapi.KvHintEntry:
    properties:
      key:
        type: string
    type: object
...
  adminapi.alertItem:
    properties:
      detail:
        type: object
      name:
        type: string
      severity:
        type: string
      since:
        type: string
      target:
        type: string
    type: object
  adminapi.amAlert:
    properties:
      annotations:
        $ref: '#/definitions/adminapi.stringMap'
      fingerprint:
        type: string
      labels:
        $ref: '#/definitions/adminapi.stringMap'
      startsAt:
        type: string
      status:
        type: string
    type: object
  adminapi.amWebhookPayload:
    properties:
      alerts:
        items:
          $ref: '#/definitions/adminapi.amAlert'
        type: array
      status:
        type: string
    type: object
...
  adminapi.defaultDef:
    properties:
      rate_limit:
        $ref: '#/definitions/adminapi.rateLimitDef'
    type: object
  adminapi.fieldDef:
    properties:
      help:
        type: string
      key:
        type: string
      kvHint:
        items:
          $ref: '#/definitions/adminapi.KvHintEntry'
        type: array
      label:
        type: string
      options:
        items:
          $ref: '#/definitions/adminapi.FieldOption'
        type: array
      required:
        type: boolean
      sensitive:
        type: boolean
      type:
        type: string
    type: object
  adminapi.formSchemaResponse:
    properties:
      schema_version:
        type: string
      vendors:
        additionalProperties:
          $ref: '#/definitions/adminapi.vendorSchemaEntry'
        type: object
    type: object
...
  adminapi.nodeDetailResponse:
    properties:
      capabilities:
        items:
          type: string
        type: array
      cold_space:
        $ref: '#/definitions/adminapi.nodeSpaceResponse'
      conn_count:
        type: integer
      healthy:
        type: boolean
      jwt:
        $ref: '#/definitions/adminapi.nodeJWTResponse'
      last_seen:
        type: string
      peer_id:
        type: string
      prefix_space:
        $ref: '#/definitions/adminapi.nodeSpaceResponse'
      recent_pin_plans:
        items:
          $ref: '#/definitions/adminapi.pinPlanRecordResponse'
        type: array
      recent_reports:
        items:
          $ref: '#/definitions/adminapi.nodeHistoryReportResponse'
        type: array
      region:
        type: string
      uptime_sec:
        type: integer
      version:
        type: string
      warm_space:
        $ref: '#/definitions/adminapi.nodeSpaceResponse'
    type: object
  adminapi.nodeHistoryReportResponse:
    properties:
      conn_count:
        type: integer
      healthy:
        type: boolean
      id:
        type: integer
      node_id:
        type: string
      peer_id:
        type: string
      prefix_total:
        type: integer
      prefix_used:
        type: integer
      received_at:
        type: string
      region:
        type: string
      reported_at:
        type: string
      version:
        type: string
      warm_total:
        type: integer
      warm_used:
        type: integer
    type: object
  adminapi.nodeJWTResponse:
    properties:
      exp:
        type: string
      should_have_renewed:
        type: boolean
    type: object
...
  adminapi.nodeSpaceResponse:
    properties:
      total:
        type: integer
      used:
        type: integer
    type: object
...
  adminapi.pinPlanRecordResponse:
    properties:
      content_id:
        type: string
      pins:
        type: integer
      sent_at:
        type: string
      seq:
        type: integer
      target_node:
        type: string
      trigger:
        type: string
      unpins:
        type: integer
    type: object
...
  adminapi.rateLimitDef:
    properties:
      burst:
        type: integer
      concurrent:
        type: integer
      qps:
        type: number
    type: object
...
  adminapi.stringMap:
    additionalProperties:
      type: string
    type: object
...
  adminapi.vendorSchemaEntry:
    properties:
      auth:
        type: string
      defaults:
        $ref: '#/definitions/adminapi.defaultDef'
      fields:
        items:
          $ref: '#/definitions/adminapi.fieldDef'
        type: array
      notes:
        items:
          type: string
        type: array
    type: object
...
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8443
info:
  contact: {}
  description: 控制面 HTTP API：节点能力 JWT 签发、blob 存储位置查询、管理后台 API 与 Prometheus 指标。
  title: MediaWorker Control-Plane API
  version: "1.0"
paths:
...
  /v1/admin/alerts:
    get:
      description: 返回当前 firing 状态的告警事件
      parameters:
      - description: 告警状态（默认 firing）
        in: query
        name: status
        type: string
      - description: 返回条数（最多 100）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/adminapi.alertItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AdminBearer: []
      summary: 活跃告警列表
      tags:
      - admin-alerts
  /v1/admin/alerts/webhook:
    post:
      consumes:
      - application/json
      description: 接收 Alertmanager v4 webhook 告警，持久化到 alert_events 表
      parameters:
      - description: Webhook 令牌
        in: header
        name: X-Alert-Token
        required: true
        type: string
      - description: Alertmanager webhook payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/adminapi.amWebhookPayload'
      produces:
      - application/json
      responses:
        "202":
          description: '{"received": N}'
          schema:
            type: object
        "400":
          description: 无效 payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: 无效 X-Alert-Token
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AlertWebhookToken: []
      summary: Alertmanager Webhook
      tags:
      - admin-alerts
...
  /v1/admin/nodes/{peer_id}:
    get:
      description: 返回指定节点的完整信息，含最近状态历史与最近派单计划
      parameters:
      - description: libp2p PeerID
        in: path
        name: peer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adminapi.nodeDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: 节点不存在
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AdminBearer: []
      summary: 节点详情
      tags:
      - admin-nodes
...
  /v1/admin/vendors/form-schema:
    get:
      description: 返回所有供应商的 auth 字段定义、默认速率限制与帮助说明
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adminapi.formSchemaResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - AdminBearer: []
      summary: 供应商表单 schema
      tags:
      - admin-vendors
...
securityDefinitions:
  AdminBearer:
    description: 管理后台 UserToken，由 POST /v1/auth/login 签发
    in: header
    name: Authorization
    type: apiKey
  AlertWebhookToken:
    description: Alertmanager webhook 令牌
    in: header
    name: X-Alert-Token
    type: apiKey
  NodeBearer:
    description: 节点能力 JWT，由 POST /v1/node/jwt 签发
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"